Accessibility fixes for the palettes. The main thing is that ARIA's "role" is relatively new and support is still buggy. I switched the relevent uses to be <ul>/<li>, etc. This required some changes to the CSS which I hope I did correctly (looks ok, but maybe there are some unintended consequences).
The palette groups have visual labels at the bottom. This is not good for
screen readers. I marked them with aria-hidden and added an SR-only
heading as the first item. This is not right as it is counted as part of
the list. I tried put it before the list with aria-labelledby pointing to
it, but it wasn't spoken. I'll work on a fix next.

There are a number of validation problems with the page, so of which were
caused by my changed. These are not addressed in this commit.