Buttons/Links in main page had different tabindex values for unknown reasons. Some where programatically focusable (<0, although really only -1 is normally used); I changed them all to be '0' so they could be tabbed to.
In the future, we might change it so only the first entry is tabable and then the rest are accessed a different way. This would make it easier for tabs to move to different parts of the interface. Currently, that can be done using jumping via lists/main, but those take a bit more discovery to know about.