Move math back into button.
Math in buttons is problematic for screen readers because they allow navigation, which in turns means key commands need to be captured. Typically, you can put something that needs mouse or key input inside another element that captures it. NVDA works fine when you do this, but not JAWS. I don't know about VoiceOver. Because JAWS won't handle this, the math is marked as aria-hidden. I'm awaiting a change to mathlive to allow some speech text to be inserted for screen readers. Currently, it just says "dummy text".