Change a <span> into a real button -- for accessibility, always use the real tag when possible.
Note: this was a link since it went to another page, but since this is a single page app, it doesn't really go to another page. Sina and I had a long discussion on this and he felt that this really should be a button. He also felt it shouldn't be a single page app, but that's a bigger issue...